Refactor to video player
Video player split into multiple components

Video player uses React signals for some of its state, this may improve performance.
